[INPLACE-356] 커서 응답시 cursorValue를 제네릭으로 확장 (#507)

+++ backend/inplace-api/src/main/java/my/inplace/api/place/PlaceController.java
@@ -83,7 +83,7 @@ public class PlaceController implements PlaceControllerApiSpec {
     public ResponseEntity<PlaceResponse.SimpleList> getPlacesByName(
         @RequestParam String placeName,
         @ModelAttribute PlaceRequest.Filter placeFilter,
-        @RequestParam(required = false) Long cursorValue,
+        @RequestParam(required = false) Double cursorValue,
         @RequestParam(required = false) Long cursorId,
         @RequestParam(defaultValue = "10") int size,
         @RequestParam(defaultValue = "score") String sort
